move optimizemodulestate
