+++ src/main/java/frc/lib/util/OptimizeModuleState.java
@@ -1,4 +1,4 @@
-package frc.lib.math;
+package frc.lib.util;
 
 import edu.wpi.first.math.geometry.Rotation2d;
 import edu.wpi.first.math.kinematics.SwerveModuleState;
